refactor: ComponentScan 으로 자동 빈 등록
