+++ src/main/java/com/programmers/voucher/CommandLineApplication.java
@@ -3,7 +3,6 @@ package com.programmers.voucher;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 
-import com.programmers.voucher.config.AppConfig;
 import com.programmers.voucher.controller.VoucherController;
 
 public class CommandLineApplication {
