+++ src/main/java/com/programmers/voucher/domain/VoucherFactory.java
@@ -2,8 +2,11 @@ package com.programmers.voucher.domain;
 
 import java.util.UUID;
 
+import org.springframework.stereotype.Component;
+
 import com.programmers.voucher.exception.WrongVoucherTypeException;
 
+@Component
 public class VoucherFactory {
 
 	public Voucher makeVoucher(String voucherType, UUID voucherId, int discountAmount) {
