+++ src/test/java/com/programmers/voucher/service/VoucherServiceTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 
-import com.programmers.voucher.config.AppConfig;
+import com.programmers.voucher.AppConfig;
 import com.programmers.voucher.domain.Voucher;
 import com.programmers.voucher.repository.VoucherRepository;
